🎨 테스트 코드 공통화.
구현 방식은 달라도 테스트 항목은 같기 때문에 공통화.
